requirements.txt
selenium
webdriver_manager
asyncio
pytest
trio
pytest-trio
flake8
python-pcapng
scapy
pandas
matplotlib
seaborn
